requirements.txt
# PSTDS v2.0
tradingagents>=0.2.0
langchain>=0.2.0
langgraph>=0.1.0
yfinance>=0.2.40
akshare>=1.14.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
pyarrow>=15.0.0
requests>=2.31.0
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.5.0
pydantic>=2.7.0
pymongo>=4.7.0
chromadb>=0.5.0
streamlit>=1.35.0
plotly>=5.22.0
apscheduler>=3.10.0
python-docx>=1.1.0
weasyprint>=62.0
markdown>=3.6
plyer>=2.1.0
pytest>=8.2.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
cryptography>=42.0.0
